added progress bar and skipped existing files

+++ duplicated_photos_remove.py
@@ -7,6 +7,7 @@ import shutil
 from PIL import Image
 import PIL
 import glob
+import tqdm
 
 source_path      = "C:\\Kornel_Zdjecia\\tmp_script_test_source_source"
 destination_path = "C:\\Kornel_Zdjecia\\tmp_script_test_source_source_helper"
@@ -141,16 +142,14 @@ class DirectoryStructure:
             [self.classify_file_for_export(file_path) for file_path in result]
 
     def export_sorted_to_destination(self, destination_path):
-        progress_val = 0
-        for export_source_path in self.trusted_files:
+        for export_source_path in tqdm.tqdm(self.trusted_files):
             export_relative_path = export_source_path.replace(self.root_directory, '')
             export_destination_path = destination_path + export_relative_path
             os.makedirs(os.path.dirname(export_destination_path), exist_ok=True)
-            shutil.copyfile(export_source_path, export_destination_path)
-            picture = Image.open(export_destination_path)
-            picture.save(export_destination_path,optimize=True,quality=30) 
-            print('Progress: {:.2f}%'.format(progress_val))
-            progress_val = self.trusted_files.index(export_source_path)/len(self.trusted_files)*100
+            if not os.path.isfile(export_destination_path):
+                shutil.copyfile(export_source_path, export_destination_path)
+                picture = Image.open(export_destination_path)
+                picture.save(export_destination_path,optimize=True,quality=30) 
 
     def find_duplicates_in_current_directory(self, full_file_path):
         if os.path.split(full_file_path)[1] in [os.path.split(f)[1] for f in self.trusted_files]:
